# internal/openapi-spec.yaml
components:
    responses:
        AnswerResponse:
            content:
                application/json:
                    schema:
                        allOf:
                            - $ref: '#/components/schemas/AnswerResult'
                            - properties:
                                costDollars:
                                    $ref: '#/components/schemas/CostDollars'
                              type: object
                text/event-stream:
                    schema:
                        properties:
                            answer:
                                description: Partial answer chunk when streaming is enabled.
                                type: string
                            citations:
                                items:
                                    $ref: '#/components/schemas/AnswerCitation'
                                type: array
                        type: object
            description: OK
        ContentsResponse:
            content:
                application/json:
                    schema:
                        properties:
                            context:
                                description: A formatted string of the search results ready for LLMs.
                                type: string
                            costDollars:
                                $ref: '#/components/schemas/CostDollars'
                            requestId:
                                description: Unique identifier for the request
                                example: e492118ccdedcba5088bfc4357a8a125
                                type: string
                            results:
                                items:
                                    $ref: '#/components/schemas/ResultWithContent'
                                type: array
                            statuses:
                                description: Status information for each requested URL
                                items:
                                    properties:
                                        error:
                                            description: Error details, only present when status is "error"
                                            nullable: true
                                            properties:
                                                httpStatusCode:
                                                    description: The corresponding HTTP status code
                                                    example: 404
                                                    nullable: true
                                                    type: integer
                                                tag:
                                                    description: Specific error type
                                                    enum:
                                                        - CRAWL_NOT_FOUND
                                                        - CRAWL_TIMEOUT
                                                        - CRAWL_LIVECRAWL_TIMEOUT
                                                        - SOURCE_NOT_AVAILABLE
                                                        - CRAWL_UNKNOWN_ERROR
                                                    example: CRAWL_NOT_FOUND
                                                    type: string
                                            type: object
                                        id:
                                            description: The URL that was requested
                                            example: https://example.com
                                            type: string
                                        status:
                                            description: Status of the content fetch operation
                                            enum:
                                                - success
                                                - error
                                            example: success
                                            type: string
                                    type: object
                                type: array
                        type: object
            description: OK
...
    schemas:
...
        AnswerResult:
            properties:
                answer:
                    description: The generated answer based on search results.
                    example: $350 billion.
                    type: string
                citations:
                    description: Search results used to generate the answer.
                    items:
                        $ref: '#/components/schemas/AnswerCitation'
                    type: array
            type: object
...
    securitySchemes:
        apikey:
            description: API key can be provided either via x-api-key header or Authorization header with Bearer scheme
            in: header
            name: x-api-key
            type: apiKey
        bearer:
            description: API key can be provided either via x-api-key header or Authorization header with Bearer scheme
            scheme: bearer
            type: http
info:
    description: A comprehensive API for internet-scale search, allowing users to perform queries and retrieve results from a wide variety of sources using embeddings-based and traditional search.
    title: Exa Search API
    version: 1.2.0
openapi: 3.1.0
paths:
    /answer:
        post:
            description: |
                Performs a search based on the query and generates either a direct answer or a detailed summary with citations, depending on the query type.
            operationId: answer
            requestBody:
                content:
                    application/json:
                        schema:
                            properties:
                                query:
                                    description: The question or query to answer.
                                    example: What is the latest valuation of SpaceX?
                                    minLength: 1
                                    type: string
                                stream:
                                    default: false
                                    description: If true, the response is returned as a server-sent events (SSS) stream.
                                    type: boolean
                                text:
                                    default: false
                                    description: If true, the response includes full text content in the search results
                                    type: boolean
                            required:
                                - query
                            type: object
                required: true
            responses:
                "200":
                    $ref: '#/components/responses/AnswerResponse'
            summary: Generate an answer from search results
            x-codeSamples:
                - label: Simple answer
                  lang: bash
                  source: |
                    curl -X POST 'https://api.exa.ai/answer' \
                      -H 'x-api-key: YOUR-EXA-API-KEY' \
                      -H 'Content-Type: application/json' \
                      -d '{
                        "query": "What is the latest valuation of SpaceX?",
                        "text": true
                      }'
                - label: Simple answer
                  lang: python
                  source: |
                    # pip install exa-py
                    from exa_py import Exa
                    exa = Exa('YOUR_EXA_API_KEY')

                    result = exa.answer(
                        "What is the latest valuation of SpaceX?",
                        text=True
                    )

                    print(result)
                - label: Simple answer
                  lang: javascript
                  source: |
                    // npm install exa-js
                    import Exa from 'exa-js';
                    const exa = new Exa('YOUR_EXA_API_KEY');

                    const result = await exa.answer(
                        'What is the latest valuation of SpaceX?',
                        { text: true }
                    );

                    console.log(result);
                - label: Simple answer
                  lang: php
                  source: ""
                - label: Simple answer
                  lang: go
                  source: ""
                - label: Simple answer
                  lang: java
                  source: ""
    /contents:
        post:
            operationId: getContents
            requestBody:
                content:
                    application/json:
                        schema:
                            allOf:
                                - properties:
                                    ids:
                                        deprecated: true
                                        description: Deprecated - use 'urls' instead. Array of document IDs obtained from searches.
                                        example:
                                            - https://arxiv.org/pdf/2307.06435
                                        items:
                                            type: string
                                        type: array
                                    urls:
                                        description: Array of URLs to crawl (backwards compatible with 'ids' parameter).
                                        example:
                                            - https://arxiv.org/pdf/2307.06435
                                        items:
                                            type: string
                                        type: array
                                  required:
                                    - urls
                                  type: object
                                - $ref: '#/components/schemas/ContentsRequest'
                            type: object
                required: true
            responses:
                "200":
                    $ref: '#/components/responses/ContentsResponse'
            summary: Get Contents
            x-codeSamples:
                - label: Simple contents retrieval
                  lang: bash
                  source: |
                    curl -X POST 'https://api.exa.ai/contents' \
                      -H 'x-api-key: YOUR-EXA-API-KEY' \
                      -H 'Content-Type: application/json' \
                      -d '{
                        "urls": ["https://arxiv.org/abs/2307.06435"],
                        "text": true
                      }'
                - label: Simple contents retrieval
                  lang: python
                  source: |
                    # pip install exa-py
                    from exa_py import Exa
                    exa = Exa('YOUR_EXA_API_KEY')

                    results = exa.get_contents(
                        urls=["https://arxiv.org/abs/2307.06435"],
                        text=True
                    )

                    print(results)
                - label: Simple contents retrieval
                  lang: javascript
                  source: |
                    // npm install exa-js
                    import Exa from 'exa-js';
                    const exa = new Exa('YOUR_EXA_API_KEY');

                    const results = await exa.getContents(
                        ["https://arxiv.org/abs/2307.06435"],
                        { text: true }
                    );

                    console.log(results);
                - label: Simple contents retrieval
                  lang: php
                  source: ""
                - label: Simple contents retrieval
                  lang: go
                  source: ""
                - label: Simple contents retrieval
                  lang: java
                  source: ""
                - label: Advanced contents retrieval
                  lang: bash
                  source: |
                    curl --request POST \
                      --url https://api.exa.ai/contents \
                      --header 'x-api-key: YOUR-EXA-API-KEY' \
                      --header 'Content-Type: application/json' \
                      --data '{
                        "urls": ["https://arxiv.org/abs/2307.06435"],
                        "text": {
                          "maxCharacters": 1000,
                          "includeHtmlTags": false
                        },
                        "highlights": {
                          "numSentences": 3,
                          "highlightsPerUrl": 2,
                          "query": "Key findings"
                        },
                        "summary": {
                          "query": "Main research contributions"
                        },
                        "subpages": 1,
                        "subpageTarget": "references",
                        "extras": {
                          "links": 2,
                          "imageLinks": 1
                        }
                      }'
                - label: Advanced contents retrieval
                  lang: python
                  source: |
                    # pip install exa-py
                    from exa_py import Exa
                    exa = Exa('YOUR_EXA_API_KEY')

                    results = exa.get_contents(
                        urls=["https://arxiv.org/abs/2307.06435"],
                        text={
                            "maxCharacters": 1000,
                            "includeHtmlTags": False
                        },
                        highlights={
                            "numSentences": 3,
                            "highlightsPerUrl": 2,
                            "query": "Key findings"
                        },
                        summary={
                            "query": "Main research contributions"
                        },
                        subpages=1,
                        subpage_target="references",
                        extras={
                            "links": 2,
                            "image_links": 1
                        }
                    )

                    print(results)
                - label: Advanced contents retrieval
                  lang: javascript
                  source: |
                    // npm install exa-js
                    import Exa from 'exa-js';
                    const exa = new Exa('YOUR_EXA_API_KEY');

                    const results = await exa.getContents(
                        ["https://arxiv.org/abs/2307.06435"],
                        {
                            text: {
                                maxCharacters: 1000,
                                includeHtmlTags: false
                            },
                            highlights: {
                                numSentences: 3,
                                highlightsPerUrl: 2,
                                query: "Key findings"
                            },
                            summary: {
                                query: "Main research contributions"
                            },
                            subpages: 1,
                            subpageTarget: "references",
                            extras: {
                                links: 2,
                                imageLinks: 1
                            }
                        }
                    );

                    console.log(results);
    /findSimilar:
        post:
            description: Find similar links to the link provided. Optionally get contents.
            operationId: findSimilar
            requestBody:
                content:
                    application/json:
                        schema:
                            allOf:
                                - properties:
                                    url:
                                        description: The url for which you would like to find similar links.
                                        example: https://arxiv.org/abs/2307.06435
                                        type: string
                                  required:
                                    - url
                                  type: object
                                - $ref: '#/components/schemas/CommonRequest'
                required: true
            responses:
                "200":
                    $ref: '#/components/responses/FindSimilarResponse'
            summary: Find similar links
            x-codeSamples:
                - label: Find similar links
                  lang: bash
                  source: |
                    curl -X POST 'https://api.exa.ai/findSimilar' \
                      -H 'x-api-key: YOUR-EXA-API-KEY' \
                      -H 'Content-Type: application/json' \
                      -d '{
                        "url": "https://arxiv.org/abs/2307.06435",
                        "text": true
                      }'
                - label: Find similar links
                  lang: python
                  source: |
                    # pip install exa-py
                    from exa_py import Exa
                    exa = Exa('YOUR_EXA_API_KEY')

                    results = exa.find_similar_and_contents(
                        url="https://arxiv.org/abs/2307.06435",
                        text=True
                    )

                    print(results)
                - label: Find similar links
                  lang: javascript
                  source: |
                    // npm install exa-js
                    import Exa from 'exa-js';
                    const exa = new Exa('YOUR_EXA_API_KEY');

                    const results = await exa.findSimilarAndContents(
                        'https://arxiv.org/abs/2307.06435',
                        { text: true }
                    );

                    console.log(results);
                - label: Find similar links
                  lang: php
                  source: ""
                - label: Find similar links
                  lang: go
                  source: ""
                - label: Find similar links
                  lang: java
                  source: ""
...
    /search:
        post:
            description: Perform a search with a Exa prompt-engineered query and retrieve a list of relevant results. Optionally get contents.
            operationId: search
            requestBody:
                content:
                    application/json:
                        schema:
                            allOf:
                                - properties:
                                    additionalQueries:
                                        description: Additional query variations for deep search. Only works with type="deep". When provided, these queries are used alongside the main query for comprehensive results.
                                        example:
                                            - LLM advancements
                                            - large language model progress
                                        items:
                                            type: string
                                        type: array
                                    category:
                                        description: A data category to focus on.
                                        enum:
                                            - company
                                            - research paper
                                            - news
                                            - pdf
                                            - github
                                            - tweet
                                            - personal site
                                            - people
                                            - financial report
                                        example: research paper
                                        type: string
                                    query:
                                        description: The query string for the search.
                                        example: Latest developments in LLM capabilities
                                        type: string
                                    type:
                                        default: auto
                                        description: The type of search. Neural uses an embeddings-based model, auto (default) intelligently combines available search methods, fast uses streamlined versions of the neural model, and deep provides comprehensive search with query expansion and detailed context.
                                        enum:
                                            - neural
                                            - fast
                                            - auto
                                            - deep
                                        example: auto
                                        type: string
                                    userLocation:
                                        description: The two-letter ISO country code of the user, e.g. US.
                                        example: US
                                        type: string
                                  required:
                                    - query
                                  type: object
                                - $ref: '#/components/schemas/CommonRequest'
                required: true
            responses:
                "200":
                    $ref: '#/components/responses/SearchResponse'
            summary: Search
            x-codeSamples:
                - label: Simple search and contents
                  lang: bash
                  source: |
                    curl -X POST 'https://api.exa.ai/search' \
                      -H 'x-api-key: YOUR-EXA-API-KEY' \
                      -H 'Content-Type: application/json' \
                      -d '{
                        "query": "Latest research in LLMs",
                        "text": true
                      }'
                - label: Simple search and contents
                  lang: python
                  source: "# pip install exa-py\nfrom exa_py import Exa\nexa = Exa('YOUR_EXA_API_KEY')\n\nresults = exa.search_and_contents(\n    \"Latest research in LLMs\", \n    text=True\n)\n\nprint(results)\n"
                - label: Simple search and contents
                  lang: javascript
                  source: "// npm install exa-js\nimport Exa from 'exa-js';\nconst exa = new Exa('YOUR_EXA_API_KEY');\n\nconst results = await exa.searchAndContents(\n    'Latest research in LLMs', \n    { text: true }\n);\n\nconsole.log(results);\n"
                - label: Simple search and contents
                  lang: php
                  source: ""
                - label: Simple search and contents
                  lang: go
                  source: ""
                - label: Simple search and contents
                  lang: java
                  source: ""
                - label: Advanced search with filters
                  lang: bash
                  source: |
                    curl --request POST \
                      --url https://api.exa.ai/search \
                      --header 'x-api-key: <token>' \
                      --header 'Content-Type: application/json' \
                      --data '{
                      "query": "Latest research in LLMs",
                      "type": "auto",
                      "category": "research paper",
                      "numResults": 10,
                      "contents": {
                        "text": true,
                        "summary": {
                          "query": "Main developments"
                        },
                        "subpages": 1,
                        "subpageTarget": "sources",
                        "extras": {
                          "links": 1,
                          "imageLinks": 1
                        }
                      }
                    }'
                - label: Advanced search with filters
                  lang: python
                  source: |
                    # pip install exa-py
                    from exa_py import Exa
                    exa = Exa('YOUR_EXA_API_KEY')

                    results = exa.search_and_contents(
                        "Latest research in LLMs",
                        type="auto",
                        category="research paper",
                        num_results=10,
                        text=True,
                        summary={
                            "query": "Main developments"
                        },
                        subpages=1,
                        subpage_target="sources",
                        extras={
                            "links": 1,
                            "image_links": 1
                        }
                    )

                    print(results)
                - label: Advanced search with filters
                  lang: javascript
                  source: |
                    // npm install exa-js
                    import Exa from 'exa-js';
                    const exa = new Exa('YOUR_EXA_API_KEY');

                    const results = await exa.searchAndContents('Latest research in LLMs', {
                        type: 'auto',
                        category: 'research paper',
                        numResults: 10,
                        contents: {
                            text: true,
                            summary: {
                                query: 'Main developments'
                            },
                            subpages: 1,
                            subpageTarget: 'sources',
                            extras: {
                                links: 1,
                                imageLinks: 1
                            }
                        }
                    });

                    console.log(results);
                - label: Deep search with query variations
                  lang: bash
                  source: |
                    curl --request POST \
                      --url https://api.exa.ai/search \
                      --header 'x-api-key: <token>' \
                      --header 'Content-Type: application/json' \
                      --data '{
                      "query": "blog post about AI",
                      "additionalQueries": ["AI blogpost", "machine learning blogs"],
                      "type": "deep",
                      "contents": {
                        "text": true,
                        "context": true
                      }
                    }'
                - label: Deep search with query variations
                  lang: python
                  source: |
                    # pip install exa-py
                    from exa_py import Exa
                    exa = Exa('YOUR_EXA_API_KEY')

                    results = exa.search_and_contents(
                        "blog post about AI",
                        type="deep",
                        additional_queries=["AI blogpost", "machine learning blogs"],
                        text=True,
                        context=True
                    )

                    print(results)
                - label: Deep search with query variations
                  lang: javascript
                  source: |
                    // npm install exa-js
                    import Exa from 'exa-js';
                    const exa = new Exa('YOUR_EXA_API_KEY');

                    const results = await exa.searchAndContents('blog post about AI', {
                        type: 'deep',
                        additionalQueries: ['AI blogpost', 'machine learning blogs'],
                        text: true,
                        context: true
                    });

                    console.log(results);
                - label: Advanced search with filters
                  lang: php
                  source: ""
                - label: Advanced search with filters
                  lang: go
                  source: ""
                - label: Advanced search with filters
                  lang: java
                  source: ""
security:
    - apikey: []
servers:
    - url: https://api.exa.ai